requirements.txt
Flask
requests
Pillow
Jinja2
audible